Add ready attribute
